LayoutTests/fast/events/spatial-navigation/snav-fully-aligned-vertically-expected.txt

test
test             test
test
test
test
test




test
test
PASS document.activeElement.getAttribute("id") is "4"
PASS document.activeElement.getAttribute("id") is "6"
PASS document.activeElement.getAttribute("id") is "8"
PASS document.activeElement.getAttribute("id") is "end"
PASS document.activeElement.getAttribute("id") is "7"
PASS document.activeElement.getAttribute("id") is "2"
